requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==4.2.8
django-celery-results==2.5.1
django-debug-toolbar==4.2.0
django-phonenumber-field==7.3.0
factory-boy==3.3.0
Faker==23.2.1
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
phonenumbers==8.13.28
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg==3.1.17
pytest==8.0.1
pytest-django==4.8.0
pytest-factoryboy==2.6.0
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
stripe==8.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
